refactor(cco): declutter ops v2 dir — inline SymmArena, move tests/bench under tests/
- merge SymmArena into dispatch_combine_op.py (drop symm_arena.py + its stale
  RoutingHandle); bench imports SymmArena from the op module.
- move test_op.py / bench_dispatch_combine.py / run_bench.sh / dist_common.py
  into tests/; fix sys.path (add parent for op+kernels) and run_bench DIR.
- README layout table + run paths updated.

No behavior change; EP8 test_op + bench still pass from tests/.
